pyproject.toml
[tool.poetry]
name = "g2f-platform"
version = "0.1.0"
description = "Enterprise Greyhound Prediction Platform"
authors = ["T <you@example.com>"]
readme = "README.md"
packages = [
    { include = "g2f_core", from = "libs" }
]

[tool.poetry.dependencies]
python = "^3.12"
# Domain & Data
pydantic = "^2.10"
pydantic-settings = "^2.6"
sqlalchemy = "^2.0"
# Ingestion
playwright = "^1.49"
# Cloud
google-cloud-storage = "^2.14"
beautifulsoup4 = "^4.14.3"
lxml = "^6.0.2"
psycopg = {extras = ["binary"], version = "^3.3.2"}
alembic = "^1.17.2"
python-dotenv = "^1.2.1"
pgvector = "^0.4.2"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-cov = "^6.0"
ruff = "^0.8"  # Replaces black, isort, flake8
mypy = "^1.13"
pre-commit = "^4.0"
types-requests = "^2.31" # Type stubs for mypy
pytest-bdd = "^8.1.0"
pytest-asyncio = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# --- RUFF CONFIGURATION (Replaces Black & Isort) ---
[tool.ruff]
line-length = 80
indent-width = 4
target-version = "py312"
exclude = [
    "ops/local_runners/data",  # Docker volumes (Permission Denied)
    "ops/db_migrations"
]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
    "E", # pycodestyle errors
    "F", # Pyflakes
    "I", # isort (Import sorting)
    "B", # flake8-bugbear (Potential bugs)
    "UP", # pyupgrade (Upgrade syntax)
    "N", # pep8-naming
    "SIM", # flake8-simplify
]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# --- MYPY CONFIGURATION ---
[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
namespace_packages = false
mypy_path = "libs"
files = [
    "apps",
    "scripts",
    "tests",
    "manual_run.py",
]
exclude = [
    '^\.venv',
    '^libs/',
    "ops/local_runners/data",
    "ops/db_migrations"
]

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "pytest_bdd.*"
]
# This tells Mypy: "Trust me, the decorators in tests are fine."
disallow_untyped_decorators = false
disallow_untyped_defs = false

[tool.pytest.ini_options]
pythonpath = [".", "libs"]
